feat: g6 react demo with webpack & babel
