--- pages/.umi/polyfills.js
@@ -1,3 +1,0 @@
-import 'core-js';
-import 'regenerator-runtime/runtime';
-
