+++ pages/index.html
@@ -1,0 +1,9 @@
+<html>
+  <head>
+      <meta charset="utf-8"/>
+      <title>G6 & React</title>
+  </head>
+  <body>
+    <div id="container"></div>
+  </body>
+</html>
